Consolidate useUserProfiles hook
Consolidate duplicated alias+avatar lookup in
- poll-comments.ts
- token-comments.ts
- roles.ts

This has to be merged into a global role check later on, all copies already shared the same SWR cache key.

rename WIKI_MODERATOR_CONTRACT -> MODERATOR_CONTRACT
