+++ src/data/messaging/poll-comments.ts
@@ -11,7 +11,6 @@ import {
   fetchBigMapValuesBulk,
 } from './api'
 import { fetchMsgIpfsJson } from './ipfs'
-import { fetchGraphQL } from '@data/api'
 import type {
   CommentPostedEvent,
   CommentPayload,
@@ -300,50 +299,3 @@ export function useIsBanned(address: string | undefined) {
     { revalidateOnFocus: false, dedupingInterval: 60_000 }
   )
 }
-
-/**
- * Batched lookup of (alias, identicon) for a set of addresses. 
- * 
- * This function will be updated with the upcoming channels branch.
- * 
- */
-interface UserProfile {
-  alias?: string
-  logo?: string
-}
-
-interface TeiaUserRow {
-  user_address: string
-  name?: string | null
-  metadata?: { data?: { identicon?: string | null } | null } | null
-}
-
-export function useUserProfiles(addresses: string[]) {
-  const unique = [...new Set((addresses ?? []).filter(Boolean))].sort()
-  return useSWR<Record<string, UserProfile>>(
-    unique.length > 0 ? `teia-user-profiles:${unique.join(',')}` : null,
-    async () => {
-      const result = await fetchGraphQL(
-        `query CommentAuthors($addresses: [String!]!) {
-           teia_users(where: { user_address: { _in: $addresses } }) {
-             user_address
-             name
-             metadata { data }
-           }
-         }`,
-        'CommentAuthors',
-        { addresses: unique }
-      )
-      const out: Record<string, UserProfile> = {}
-      for (const row of (result?.data?.teia_users ?? []) as TeiaUserRow[]) {
-        out[row.user_address] = {
-          alias: row.name ?? undefined,
-          logo: row.metadata?.data?.identicon ?? undefined,
-        }
-      }
-      return out
-    },
-    { revalidateOnFocus: false, dedupingInterval: 60_000 }
-  )
-}
-
