+++ src/data/messaging/token-comments.ts
@@ -277,49 +277,3 @@ export function useIsBanned(address: string | undefined) {
     { revalidateOnFocus: false, dedupingInterval: 60_000 }
   )
 }
-
-/**
- * Batched lookup of (alias, identicon) for a set of addresses.
- *
- * Same shape as poll-comments useUserProfiles, will move to a shared
- * location when we extract the messaging UI.
- */
-interface UserProfile {
-  alias?: string
-  logo?: string
-}
-
-interface TeiaUserRow {
-  user_address: string
-  name?: string | null
-  metadata?: { data?: { identicon?: string | null } | null } | null
-}
-
-export function useUserProfiles(addresses: string[]) {
-  const unique = [...new Set((addresses ?? []).filter(Boolean))].sort()
-  return useSWR<Record<string, UserProfile>>(
-    unique.length > 0 ? `teia-user-profiles:${unique.join(',')}` : null,
-    async () => {
-      const result = await fetchGraphQL(
-        `query CommentAuthors($addresses: [String!]!) {
-           teia_users(where: { user_address: { _in: $addresses } }) {
-             user_address
-             name
-             metadata { data }
-           }
-         }`,
-        'CommentAuthors',
-        { addresses: unique }
-      )
-      const out: Record<string, UserProfile> = {}
-      for (const row of (result?.data?.teia_users ?? []) as TeiaUserRow[]) {
-        out[row.user_address] = {
-          alias: row.name ?? undefined,
-          logo: row.metadata?.data?.identicon ?? undefined,
-        }
-      }
-      return out
-    },
-    { revalidateOnFocus: false, dedupingInterval: 60_000 }
-  )
-}
