+++ src/pages/wiki/WikiAdmin.jsx
@@ -3,7 +3,7 @@ import { useOutletContext, Navigate, Link } from 'react-router-dom'
 import { Button } from '@atoms/button'
 import { PATH } from '@constants'
 import { walletPreview } from '@utils/string'
-import { useUserProfiles } from '@data/messaging/token-comments'
+import { useUserProfiles } from '@data/roles'
 import { setPageHidden } from '@data/wiki'
 import WikiProposalList from './WikiProposalList'
 import styles from '@style'
