+++ src/pages/wiki/WikiHistory.jsx
@@ -4,7 +4,7 @@ import { Button } from '@atoms/button'
 import { PATH } from '@constants'
 import { walletPreview } from '@utils/string'
 import { msgIpfsToUrl } from '@data/messaging/ipfs'
-import { useUserProfiles } from '@data/messaging/token-comments'
+import { useUserProfiles } from '@data/roles'
 import { useWikiVersions, useWikiPageContent } from '@data/wiki'
 import { WikiMarkdown } from '@components/wiki'
 import styles from '@style'
