+++ src/pages/wiki/WikiPage.jsx
@@ -3,7 +3,7 @@ import { useOutletContext, useParams, Link, Navigate } from 'react-router-dom'
 import { Button } from '@atoms/button'
 import { PATH } from '@constants'
 import { walletPreview } from '@utils/string'
-import { useUserProfiles } from '@data/messaging/token-comments'
+import { useUserProfiles } from '@data/roles'
 import { useWikiPageContent, setPageHidden } from '@data/wiki'
 import { WikiMarkdown } from '@components/wiki'
 import styles from '@style'
